+++ tests/run-almost-all.ts
@@ -134,6 +134,7 @@ import { testit as issue6037 } from "./issue-6037.ts";
 import { testit as issue6040 } from "./issue-6040.ts";
 import { testit as issue6043 } from "./issue-6043.ts";
 import { testit as issue6045 } from "./issue-6045.ts";
+import { testit as issue6046 } from "./issue-6046.ts";
 import { testit as favoritesTabAfterDocumentClose } from "./favorites-tab-after-document-close.ts";
 import { testit as movePolygonPolylineInk } from "./move-polygon-polyline-ink.ts";
 import { testit as annotDeleteRedraw } from "./annot-delete-redraw.ts";
@@ -221,6 +222,7 @@ export const tests: NamedTest[] = [
   ["issue-6037", issue6037],
   ["issue-6040", issue6040],
   ["issue-6043", issue6043],
+  ["issue-6046", issue6046],
   ["move-polygon-polyline-ink", movePolygonPolylineInk],
   ["annot-delete-redraw", annotDeleteRedraw],
   ["annot-icon-over-contents", annotIconOverContents],